I think this naming makes more sense. Also move reporting code into a shared location
