+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Diagnostics/VSDocumentDiagnosticsEndpoint.cs
@@ -38,7 +38,7 @@ internal class VSDocumentDiagnosticsEndpoint(
     private readonly RazorTranslateDiagnosticsService _translateDiagnosticsService = translateDiagnosticsService;
     private readonly RazorLSPOptionsMonitor _razorLSPOptionsMonitor = razorLSPOptionsMonitor;
     private readonly ITelemetryReporter? _telemetryReporter = telemetryReporter;
-    private ImmutableDictionary<ProjectKey, int> _lastReportedProjectTagHelperCount = ImmutableDictionary<ProjectKey, int>.Empty;
+    private readonly MissingTagHelperTelemetryReporter? _missingTagHelperTelemetryReporter = telemetryReporter is null ? null : new(telemetryReporter);
 
     public bool MutatesSolutionState => false;
 
@@ -91,9 +91,12 @@ internal class VSDocumentDiagnosticsEndpoint(
         using var __ = _telemetryReporter?.TrackLspRequest(VSInternalMethods.DocumentPullDiagnosticName, LanguageServerConstants.RazorLanguageServerName, TelemetryThresholds.DiagnosticsRazorTelemetryThreshold, correlationId);
 
         var documentSnapshot = documentContext.Snapshot;
-        var razorDiagnostics = await GetRazorDiagnosticsAsync(documentSnapshot, cancellationToken).ConfigureAwait(false);
+        var razorDiagnostics = await RazorDiagnosticHelper.GetRazorDiagnosticsAsync(documentSnapshot, cancellationToken).ConfigureAwait(false);
 
-        await ReportRZ10012TelemetryAsync(documentContext, razorDiagnostics, cancellationToken).ConfigureAwait(false);
+        if (_missingTagHelperTelemetryReporter is not null && razorDiagnostics is not null)
+        {
+            await _missingTagHelperTelemetryReporter.ReportRZ10012TelemetryAsync(documentContext, razorDiagnostics, cancellationToken).ConfigureAwait(false);
+        }
 
         var (csharpDiagnostics, htmlDiagnostics) = await GetHtmlCSharpDiagnosticsAsync(documentContext, correlationId, cancellationToken).ConfigureAwait(false);
 
@@ -108,7 +111,11 @@ internal class VSDocumentDiagnosticsEndpoint(
         if (razorDiagnostics is not null)
         {
             // No extra work to do for Razor diagnostics
-            allDiagnostics.AddRange(razorDiagnostics);
+            allDiagnostics.Add(new VSInternalDiagnosticReport()
+            {
+                Diagnostics = razorDiagnostics,
+                ResultId = Guid.NewGuid().ToString()
+            });
         }
 
         if (csharpDiagnostics is not null)
@@ -146,30 +153,6 @@ internal class VSDocumentDiagnosticsEndpoint(
         return allDiagnostics.ToArray();
     }
 
-    private static async Task<VSInternalDiagnosticReport[]?> GetRazorDiagnosticsAsync(IDocumentSnapshot documentSnapshot, CancellationToken cancellationToken)
-    {
-        var codeDocument = await documentSnapshot.GetGeneratedOutputAsync(cancellationToken).ConfigureAwait(false);
-        var sourceText = codeDocument.Source.Text;
-        var csharpDocument = codeDocument.GetCSharpDocument();
-        var diagnostics = csharpDocument.Diagnostics;
-
-        if (diagnostics.Length == 0)
-        {
-            return null;
-        }
-
-        var convertedDiagnostics = RazorDiagnosticConverter.Convert(diagnostics, sourceText, documentSnapshot);
-
-        return
-        [
-            new()
-            {
-                Diagnostics = convertedDiagnostics,
-                ResultId = Guid.NewGuid().ToString()
-            }
-        ];
-    }
-
     private async Task<(VSInternalDiagnosticReport[]? CSharpDiagnostics, VSInternalDiagnosticReport[]? HtmlDiagnostics)> GetHtmlCSharpDiagnosticsAsync(DocumentContext documentContext, Guid correlationId, CancellationToken cancellationToken)
     {
         var delegatedParams = new DelegatedDiagnosticParams(documentContext.GetTextDocumentIdentifierAndVersion(), correlationId);
@@ -184,60 +167,5 @@ internal class VSDocumentDiagnosticsEndpoint(
         }
 
         return (delegatedResponse.CSharpDiagnostics, delegatedResponse.HtmlDiagnostics);
-    }
-
-    /// <summary>
-    /// Reports telemetry for RZ10012 "Found markup element with unexpected name" to help track down potential issues
-    /// with taghelpers being discovered (or lack thereof)
-    /// </summary>
-    private async ValueTask ReportRZ10012TelemetryAsync(DocumentContext documentContext, VSInternalDiagnosticReport[]? razorDiagnostics, CancellationToken cancellationToken)
-    {
-        if (razorDiagnostics is null)
-        {
-            return;
-        }
-
-        if (_telemetryReporter is null)
-        {
-            return;
-        }
-
-        var relevantDiagnosticsCount = razorDiagnostics.Sum(CountDiagnostics);
-        if (relevantDiagnosticsCount == 0)
-        {
-            return;
-        }
-
-        var tagHelpers = await documentContext.Project.GetTagHelpersAsync(cancellationToken).ConfigureAwait(false);
-        var tagHelperCount = tagHelpers.Length;
-        var shouldReport = false;
-
-        ImmutableInterlocked.AddOrUpdate(
-            ref _lastReportedProjectTagHelperCount,
-            documentContext.Project.Key,
-            (k) =>
-            {
-                shouldReport = true;
-                return tagHelperCount;
-            },
-            (k, currentValue) =>
-            {
-                shouldReport = currentValue != tagHelperCount;
-                return tagHelperCount;
-            });
-
-        if (shouldReport)
-        {
-            _telemetryReporter.ReportEvent(
-                "RZ10012",
-                Severity.Low,
-                new("tagHelpers", tagHelperCount),
-                new("RZ10012errors", relevantDiagnosticsCount),
-                new("project", documentContext.Project.Key.Id));
-        }
-
-        static int CountDiagnostics(VSInternalDiagnosticReport report)
-            => report.Diagnostics?.Count(d => d.Code == ComponentDiagnosticFactory.UnexpectedMarkupElement.Id)
-            ?? 0;
     }
 }
