+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Diagnostics/RazorDiagnosticConverterTest.cs
@@ -24,7 +24,7 @@ public class RazorDiagnosticConverterTest(ITestOutputHelper testOutput) : Langua
         var sourceText = SourceText.From(string.Empty);
 
         // Act
-        var diagnostic = RazorDiagnosticConverter.Convert(razorDiagnostic, sourceText, documentSnapshot: null);
+        var diagnostic = RazorDiagnosticHelper.ConvertToVSDiagnostic(razorDiagnostic, sourceText, documentSnapshot: null);
 
         // Assert
         Assert.Equal(razorDiagnostic.Id, diagnostic.Code);
@@ -40,7 +40,7 @@ public class RazorDiagnosticConverterTest(ITestOutputHelper testOutput) : Langua
         var expectedSeverity = DiagnosticSeverity.Error;
 
         // Act
-        var severity = RazorDiagnosticConverter.ConvertSeverity(RazorDiagnosticSeverity.Error);
+        var severity = RazorDiagnosticHelper.ConvertSeverity(RazorDiagnosticSeverity.Error);
 
         // Assert
         Assert.Equal(expectedSeverity, severity);
@@ -53,7 +53,7 @@ public class RazorDiagnosticConverterTest(ITestOutputHelper testOutput) : Langua
         var expectedSeverity = DiagnosticSeverity.Warning;
 
         // Act
-        var severity = RazorDiagnosticConverter.ConvertSeverity(RazorDiagnosticSeverity.Warning);
+        var severity = RazorDiagnosticHelper.ConvertSeverity(RazorDiagnosticSeverity.Warning);
 
         // Assert
         Assert.Equal(expectedSeverity, severity);
@@ -68,7 +68,7 @@ public class RazorDiagnosticConverterTest(ITestOutputHelper testOutput) : Langua
         var expectedRange = VsLspFactory.CreateSingleLineRange(line: 0, character: 3, length: 4);
 
         // Act
-        var range = RazorDiagnosticConverter.ConvertSpanToRange(sourceSpan, sourceText);
+        var range = RazorDiagnosticHelper.ConvertSpanToRange(sourceSpan, sourceText);
 
         // Assert
         Assert.Equal("lo W", sourceText.GetSubTextString(sourceText.GetTextSpan(range)));
@@ -84,7 +84,7 @@ public class RazorDiagnosticConverterTest(ITestOutputHelper testOutput) : Langua
         var expectedRange = VsLspFactory.DefaultRange;
 
         // Act
-        var range = RazorDiagnosticConverter.ConvertSpanToRange(sourceSpan, sourceText);
+        var range = RazorDiagnosticHelper.ConvertSpanToRange(sourceSpan, sourceText);
 
         // Assert
         Assert.Equal("", sourceText.GetSubTextString(sourceText.GetTextSpan(range)));
@@ -100,7 +100,7 @@ public class RazorDiagnosticConverterTest(ITestOutputHelper testOutput) : Langua
         var expectedRange = VsLspFactory.CreateZeroWidthRange(0, 11);
 
         // Act
-        var range = RazorDiagnosticConverter.ConvertSpanToRange(sourceSpan, sourceText);
+        var range = RazorDiagnosticHelper.ConvertSpanToRange(sourceSpan, sourceText);
 
         // Assert
         Assert.Equal("", sourceText.GetSubTextString(sourceText.GetTextSpan(range)));
@@ -116,7 +116,7 @@ public class RazorDiagnosticConverterTest(ITestOutputHelper testOutput) : Langua
         var expectedRange = VsLspFactory.CreateSingleLineRange(line: 0, character: 6, length: 5);
 
         // Act
-        var range = RazorDiagnosticConverter.ConvertSpanToRange(sourceSpan, sourceText);
+        var range = RazorDiagnosticHelper.ConvertSpanToRange(sourceSpan, sourceText);
 
         // Assert
         Assert.Equal("World", sourceText.GetSubTextString(sourceText.GetTextSpan(range)));
@@ -131,7 +131,7 @@ public class RazorDiagnosticConverterTest(ITestOutputHelper testOutput) : Langua
         var sourceText = SourceText.From(string.Empty);
 
         // Act
-        var range = RazorDiagnosticConverter.ConvertSpanToRange(sourceSpan, sourceText);
+        var range = RazorDiagnosticHelper.ConvertSpanToRange(sourceSpan, sourceText);
 
         // Assert
         Assert.Null(range);
